♻️ refactor: 테스트코드 위치 변경
